Add PRADA attacks to blackbox attack

+++ mef/utils/pytorch/lighting/module.py
@@ -41,7 +41,8 @@ class _MefModel(pl.LightningModule, ABC):
     ) -> torch.Tensor:
         x, y = batch
 
-        preds = self._shared_step_output(x).float()
+        preds = self._shared_step_output(x)
+
 
         # preds is expected to in shape of [B] for binary and [B, C] for
         # multiclass
